# Chunked internal nodes