--- a/include/pstore/adt/chunked_vector.hpp
+++ b/include/pstore/adt/chunked_vector.hpp
@@ -365,6 +365,7 @@ namespace pstore {
             --index_;
         } else {
             --it_;
+            assert (it_->size () > 0);
             index_ = it_->size () - 1U;
         }
         return *this;