… lib/core/hamt_map_types.cpp
… unittests/adt/test_chunked_vector.cpp
… unittests/core/test_hamt_map.cpp