Internal index nodes are now allocated using a `chunked_vector<>`. This performs the job of a memory pool: allocating memory in lumps sufficient for `internal_nodes_per_chunk` entries which are then consumed as new in-heap internal nodes are created.

The effect of this is to significantly reduce the number of memory allocations performed by the index code under heavy insertion pressure as thus give us a little extra performance. The cost is that we may waste as much as one chunk's worth of memory [minus `sizeof (internal_node)`].

As of this change, we’re still allocating nodes at their maximum size. That is, if we need an internal-node with enough room for 2 entries, it is currently allocated with space for 64. That’s clearly wasteful, but it simplifies the management of this records since the number of entries may grow as further insertions as carried out. Introducing an allocation pool into this process is a good first step towards having multiple pools for internal nodes with differing numbers of entries.

The point of this is to improve performance, and I do know how we love statistics. So…

Using the command:

    rld-gen \
        --external 50000 \
        --linkonce 50000 \
        --modules 100 \
        --output-directory ~/gen/output \
        --repo ~/gen/output/repository.db \
        --triple x86_64-pc-linux-gnu-repo

This will work the index code pretty hard to create:

- 5,050,000 entries in the fragment index (50,000 externals for each of the 100 compilations and 50,000 link-once entries shared amongst them).
- 5,050,002 entries in the names index. That’s 1 name per fragment, 1 for the triple, 1 for the ticket file path record. The latter two are the same for all the compilations.
- 100 entries in the compilation index (1 for each ticket file).

I captured the following for this process:

|                          | Before    | After    |
| ------------------------ | --------- | -------- |
| Run-time                 | 15.7 s    | 12.9 s   |
| No. of heap allocations  | 1,775,429 | 6,379    |
| Heap consumption         | 1.59 GiB  | 2.49 GiB |

That an ~18% reduction in run-time. Given that [Timey](https://github.com/paulhuggett/timey/), by default, runs [rld-gen](https://github.com/SNSystems/rld/tree/master/tools/gen) 2,500 times, this should significantly reduce the time to capture a set of results.

It’s worth noting that although improving rld-gen’s performance was the motivation and test-harness for this change, *any* pstore producer should see some benefit (including the compiler).